requirements.txt
spacy>=3.7.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.1.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
apscheduler>=3.10.0
fastapi>=0.104.0
uvicorn>=0.24.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
python-dateutil>=2.8.2
